\name{mlrMBO-service-package}
\alias{mlrMBO-service-package}
\alias{mlrMBO-service}
\docType{package}
\title{
\packageTitle{mlrMBO-service}
}
\description{
\packageDescription{mlrMBO-service}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{mlrMBO-service}
\packageIndices{mlrMBO-service}
~~ An overview of how to use the package, including the most ~~
~~ important functions ~~
}
\author{
\packageAuthor{mlrMBO-service}

Maintainer: \packageMaintainer{mlrMBO-service}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
